Refactor Reflector To Allow For Reading MAC Address
Certain TLVs require that the reflector know the source MAC address of
the packet in which the TLV was received. With the Reflector's prior
architecture, accessing the sender's packet's source MAC address was not
possible. This new architecture allows for that, but does introduce some
complications (the biggest of which is seeing packets for a server on
the localhost [from a client on localhost] twice). I hope to be able to
fix that in the future.

TODO: Testing and other (minimal) code cleanup are still required.

Signed-off-by: Will Hawkins <hawkinsw@obs.cr>