New Net Configuration System For Handlers
The system for TLV handlers to make changes to the header of a reflected
packet was ad hoc and required duplicated code. A new system centralizes
changes to the IP header of the reflected packet while still allowing
TLV handlers to make the necessary changes to the IP header.

With the change, the TlvHandler trait is reconfigured.

TODO: Additional documentation on what a TlvHandler should do in each of
the trait methods.

Signed-off-by: Will Hawkins <hawkinsw@obs.cr>