Add Access to SessionData For Handlers
It's possible that a TlvHandler will want to modify the session data
during its work. This change makes that possible.

It does require that the handler have exclusive access to the sessions
while processing. It's possible that, in the future, we could lock
individual sessions rather than the entire group of sessions.

Signed-off-by: Will Hawkins <hawkinsw@obs.cr>